Convert spreadsheetview to tsx
